// src/common/enums/permission.enum.ts
export enum SystemModule {
  PROJECTS = 'PROJECTS',
  SERVICES = 'SERVICES',
  BLOG = 'BLOG',
  CONTACT = 'CONTACT',
  CMS = 'CMS',
  MEDIA = 'MEDIA',
  ROLES = 'ROLES',
  USERS = 'USERS',
  LOGS = 'LOGS',
}

export enum PermissionAction {
  VIEW = 'canView',
  CREATE = 'canCreate',
  UPDATE = 'canUpdate',
  DELETE = 'canDelete',
}